refactor: clean some unused value
clean some unused value

Nddtfjiang <zhicheng.jiang@merico.dev>
