+++ plugins/gitlab/tasks/gitlab_api_commit_collector.go
@@ -2,14 +2,10 @@ package tasks
 
 import (
 	"github.com/merico-dev/lake/plugins/core"
-	"github.com/merico-dev/lake/plugins/gitlab/models"
 	"github.com/merico-dev/lake/plugins/helper"
 )
 
 const RAW_COMMIT_TABLE = "gitlab_api_commit"
-
-type ApiCommitResponse []GitlabApiCommit
-type ApiMergeRequestCommitResponse []GitlabApiCommit
 
 type GitlabApiCommit struct {
 	GitlabId       string `json:"id"`
@@ -50,25 +46,4 @@ func CollectApiCommits(taskCtx core.SubTaskContext) error {
 	}
 
 	return collector.Execute()
-}
-
-// Convert the API response to our DB model instance
-func ConvertCommit(commit *GitlabApiCommit) (*models.GitlabCommit, error) {
-	gitlabCommit := &models.GitlabCommit{
-		Sha:            commit.GitlabId,
-		Title:          commit.Title,
-		Message:        commit.Message,
-		ShortId:        commit.ShortId,
-		AuthorName:     commit.AuthorName,
-		AuthorEmail:    commit.AuthorEmail,
-		AuthoredDate:   commit.AuthoredDate.ToTime(),
-		CommitterName:  commit.CommitterName,
-		CommitterEmail: commit.CommitterEmail,
-		CommittedDate:  commit.CommittedDate.ToTime(),
-		WebUrl:         commit.WebUrl,
-		Additions:      commit.Stats.Additions,
-		Deletions:      commit.Stats.Deletions,
-		Total:          commit.Stats.Total,
-	}
-	return gitlabCommit, nil
 }
